Moved model files, started sequelize migration setup and some more
